CR - Refactor
