+++ src/app/_components/TextLink/TextLink.module.scss
@@ -1,0 +1,3 @@
+.base {
+  @apply text-brand-300 focus:brand-outline hover:underline;
+}
