+++ src/app/ecosystem-explorer/[slug]/page.tsx
@@ -22,7 +22,8 @@ import { MarkdownContent } from '@/components/MarkdownContent'
 import { ShareArticle } from '@/components/ShareArticle'
 import { StructuredDataScript } from '@/components/StructuredDataScript'
 import { TagLabel } from '@/components/TagLabel'
-import { ExternalTextLink, SmartTextLink } from '@/components/TextLink'
+import { ExternalTextLink } from '@/components/TextLink/ExternalTextLink'
+import { SmartTextLink } from '@/components/TextLink/SmartTextLink'
 import { YouTubeVideoEmbed } from '@/components/YouTubeVideoEmbed'
 
 import { getEcosystemProjectData } from '../utils/getEcosystemProjectData'
