+++ src/app/events/[slug]/components/ScheduleSection/EventDetails.tsx
@@ -2,7 +2,7 @@
 
 import { BasicCard } from '@/components/BasicCard'
 import { Heading } from '@/components/Heading'
-import { SmartTextLink } from '@/components/TextLink'
+import { SmartTextLink } from '@/components/TextLink/SmartTextLink'
 
 import type { Event } from '../../../schemas/ScheduleSchema'
 import { formatTime } from '../../utils/dateUtils'
