+++ src/app/filecoin-plus/allocators/data/allocatorsTableColumnsData.tsx
@@ -1,7 +1,7 @@
 import { createColumnHelper } from '@tanstack/react-table'
 
 import { TooltipHeader } from '@/components/Table/TooltipHeader'
-import { ExternalTextLink } from '@/components/TextLink'
+import { ExternalTextLink } from '@/components/TextLink/ExternalTextLink'
 
 import type { Allocator } from '../schemas/AllocatorSchema'
