+++ src/app/governance/components/CTAPageSection.tsx
@@ -1,7 +1,7 @@
 import { FILECOIN_FOUNDATION_URLS } from '@/constants/siteMetadata'
 
 import { CTASection } from '@/components/CTASection'
-import { ExternalTextLink } from '@/components/TextLink'
+import { ExternalTextLink } from '@/components/TextLink/ExternalTextLink'
 
 export function CTAPageSection() {
   return (
