+++ src/app/grants/data/applicationProcessData.tsx
@@ -1,4 +1,4 @@
-import { ExternalTextLink } from '@/components/TextLink'
+import { ExternalTextLink } from '@/components/TextLink/ExternalTextLink'
 
 export const applicationProcessData = [
   {
