+++ src/app/grants/data/submissionCriteriaData.tsx
@@ -1,6 +1,6 @@
 import { Code, BracketsAngle, Users } from '@phosphor-icons/react/dist/ssr'
 
-import { ExternalTextLink } from '@/components/TextLink'
+import { ExternalTextLink } from '@/components/TextLink/ExternalTextLink'
 
 export const submissionCriteriaData = [
   {
