+++ src/app/orbit/data/ambassadorsData.tsx
@@ -1,6 +1,6 @@
 import type { StaticImageProps } from '@/types/imageType'
 
-import { ExternalTextLink } from '@/components/TextLink'
+import { ExternalTextLink } from '@/components/TextLink/ExternalTextLink'
 
 import austinImg from '@/assets/orbit/ambassadors/austin-web3-filecoin-meetup.jpg'
 import barcelonaImg from '@/assets/orbit/ambassadors/barcelona-web3fc-event.jpg'
