+++ src/app/security/data/developerSupportData.tsx
@@ -5,7 +5,7 @@ import { FILECOIN_FOUNDATION_URLS } from '@/constants/siteMetadata'
 import { extractEmailAddress } from '@/utils/extractEmailAddress'
 
 import type { IconProps } from '@/components/Icon'
-import { ExternalTextLink } from '@/components/TextLink'
+import { ExternalTextLink } from '@/components/TextLink/ExternalTextLink'
 
 type DeveloperSupportData = {
   heading: {
